backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.7.0
pymongo==4.10.1
pydantic==2.10.4
pydantic-settings==2.7.1
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
APScheduler==3.10.4
email-validator==2.2.0
